refactor: Written and relocated file to examples
